cognitive_service/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
google-generativeai
python-dotenv
pydantic